+++ sql/migrations/2.3-19-natures_dossiers.sql
@@ -1,0 +1,6 @@
+SET NAMES 'utf8';
+
+# Nouvelles natures de dossier
+INSERT INTO `dossiernatureliste` (`ID_DOSSIERNATURE`, `LIBELLE_DOSSIERNATURE`, `ID_DOSSIERTYPE`, `ORDRE`) VALUES (61,"Autorisation d'une ICPE",1,23);
+INSERT INTO `dossiernatureliste` (`ID_DOSSIERNATURE`, `LIBELLE_DOSSIERNATURE`, `ID_DOSSIERTYPE`, `ORDRE`) VALUES (62,"Certificats d'urbanisme (CU)",1,24);
+INSERT INTO `dossiernatureliste` (`ID_DOSSIERNATURE`, `LIBELLE_DOSSIERNATURE`, `ID_DOSSIERTYPE`, `ORDRE`) VALUES (63,"Permis d'aménager (PA)",1,25);